+++ port_ocean/context/event.py
@@ -1,21 +1,37 @@
+import asyncio
 from contextlib import asynccontextmanager
 from dataclasses import dataclass, field
-from typing import AsyncIterator, Literal, Any, TYPE_CHECKING, Optional
+from typing import (
+    AsyncIterator,
+    Literal,
+    Any,
+    TYPE_CHECKING,
+    Optional,
+    Callable,
+    Awaitable,
+    Union,
+)
 from uuid import uuid4
 
 from loguru import logger
+from pydispatch import dispatcher  # type: ignore
 from werkzeug.local import LocalStack, LocalProxy
 
-from port_ocean.exceptions.context import EventContextNotFoundError
+from port_ocean.context.resource import resource
+from port_ocean.exceptions.context import (
+    EventContextNotFoundError,
+    ResourceContextNotFoundError,
+)
 from port_ocean.utils import get_time
 
 if TYPE_CHECKING:
     from port_ocean.core.handlers.port_app_config.models import (
-        PortAppConfig,
         ResourceConfig,
+        PortAppConfig,
     )
 
 TriggerType = Literal["manual", "machine", "request"]
+AbortCallbackFunction = Callable[[], Union[Any, Awaitable[Any]]]
 
 
 class EventType:
@@ -28,11 +44,39 @@ class EventType:
 class EventContext:
     event_type: str
     trigger_type: TriggerType = "machine"
-    resource_config: Optional["ResourceConfig"] = None
     attributes: dict[str, Any] = field(default_factory=dict)
+    _aborted: bool = False
     _port_app_config: Optional["PortAppConfig"] = None
     _parent_event: Optional["EventContext"] = None
     _event_id: str = field(default_factory=lambda: str(uuid4()))
+    _on_abort_callbacks: list[AbortCallbackFunction] = field(default_factory=list)
+
+    def on_abort(self, func: AbortCallbackFunction) -> None:
+        self._on_abort_callbacks.append(func)
+
+    def abort(self) -> None:
+        for func in self._on_abort_callbacks:
+            try:
+                if asyncio.iscoroutinefunction(func):
+                    asyncio.get_running_loop().run_until_complete(func())
+                else:
+                    func()
+            except Exception as ex:
+                logger.warning(
+                    f"Failed to call one of the abort callbacks {ex}", exc_info=True
+                )
+        self._aborted = True
+
+    @property
+    def aborted(self) -> bool:
+        return self._aborted
+
+    @property
+    def resource_config(self) -> Optional["ResourceConfig"]:
+        try:
+            return resource.resource_config
+        except ResourceContextNotFoundError:
+            return None
 
     @property
     def id(self) -> str:
@@ -80,24 +124,31 @@ event: EventContext = LocalProxy(lambda: _get_event_context())  # type: ignore
 async def event_context(
     event_type: str,
     trigger_type: TriggerType = "manual",
-    resource_config: Optional["ResourceConfig"] = None,
     attributes: dict[str, Any] | None = None,
 ) -> AsyncIterator[EventContext]:
     attributes = attributes or {}
 
     parent = _event_context_stack.top
-
-    _event_context_stack.push(
-        EventContext(
-            event_type,
-            trigger_type=trigger_type,
-            resource_config=resource_config,
-            attributes=attributes,
-            _parent_event=parent,
-            # inherit port app config from parent event so it can be used in nested events
-            _port_app_config=parent.port_app_config if parent else None,
-        )
+    new_event = EventContext(
+        event_type,
+        trigger_type=trigger_type,
+        attributes=attributes,
+        _parent_event=parent,
+        # inherit port app config from parent event, so it can be used in nested events
+        _port_app_config=parent.port_app_config if parent else None,
     )
+    _event_context_stack.push(new_event)
+
+    def _handle_event(triggering_event_id: int) -> None:
+        if (
+            new_event.event_type == EventType.RESYNC
+            and new_event.id != triggering_event_id
+        ):
+            logger.warning("ABORTING RESYNC EVENT DUE TO NEWER RESYNC REQUEST")
+            new_event.abort()
+
+    dispatcher.connect(_handle_event, event_type)
+    dispatcher.send(event_type, triggering_event_id=event.id)
 
     start_time = get_time(seconds_precision=False)
     with logger.contextualize(
@@ -124,5 +175,7 @@ async def event_context(
                 success=success,
                 time_elapsed=time_elapsed,
             ).info("Event finished")
+
+            dispatcher.disconnect(_handle_event, event_type)
 
     _event_context_stack.pop()